requirements.txt
streamlit==1.31.0
pandas==2.2.0
groq==0.4.2
tavily-python==0.3.3
requests==2.31.0
python-dotenv==1.0.0
